moved button interaction handling to events/
